feat(contracts-communication): use uint64 for chain IDs and DB nonces (#2453)
* Update chainId constants to be uint64

* Update `srcChainId` references

* Update `dstChainId` references

* Update `chainId` references

* Update `remoteChainId` references

* Fix MessageBus w/o breaking interface

* Fix other contracts

* Update nonce constants to be uint64

* Update `dbNonce` to be uint64

* Update tests

* Add InterchainTx Header compression

* Use compact encoding for ic transactions

* Add InterchainBatch Key compression

* Use compact encoding for batches

* Use new BatchKey type for DB mapping

* Cleanup unused/deprecated functions

* Go changes for #2453 (#2454)

* Regenerate `committee`

* Fix committee E2E test

* Regenerate `sin-executor`

* Fix broken casts in `sin-executor`

* Fix sin-executor tests

* Update PingPong integration tests to use uint64

* Revert "Regenerate `sin-executor`"

This reverts commit 01fa75c9b8dacfd695c920515c8ff631f3770b36.

* Revert "Regenerate `committee`"

This reverts commit e096bb2e07df72f409045d9b7c75bdfc0c349e7f.

* Fix: forgotten uint256 indexed chainId fields

* Regenerate `committee`

* Update `committee` with uint256 -> uint64 change

* Regenerate `sin-executor`

+++ packages/contracts-communication/contracts/events/InterchainAppV1Events.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.0;
 
 abstract contract InterchainAppV1Events {
     event AppConfigV1Set(uint256 requiredResponses, uint256 optimisticPeriod);
-    event AppLinked(uint256 indexed chainId, bytes32 indexed remoteApp);
+    event AppLinked(uint64 indexed chainId, bytes32 indexed remoteApp);
     event ExecutionServiceSet(address executionService);
     event TrustedModuleAdded(address module);
     event TrustedModuleRemoved(address module);
